Move string_wrapper into separate file
