+++ server/svix-server/src/core/types/mod.rs
@@ -15,9 +15,10 @@ use std::ops::Deref;
 use svix_ksuid::*;
 use validator::{Validate, ValidationErrors};
 
-use crate::{err_generic, v1::utils::validation_error};
+use crate::{err_generic, string_wrapper, v1::utils::validation_error};
 
 pub mod metadata;
+pub mod strings;
 
 use super::cryptography::{AsymmetricKey, Encryption};
 
@@ -242,76 +243,9 @@ pub trait BaseUid: Deref<Target = String> {
     }
 }
 
-macro_rules! string_wrapper {
-    ($name_id:ident) => {
-        #[derive(Clone, Debug, Hash, Eq, PartialEq, Serialize, Deserialize)]
-        pub struct $name_id(pub String);
-
-        impl Deref for $name_id {
-            type Target = String;
-
-            fn deref(&self) -> &Self::Target {
-                &self.0
-            }
-        }
-
-        impl From<$name_id> for sea_orm::Value {
-            fn from(v: $name_id) -> Self {
-                Self::String(Some(Box::new(v.0)))
-            }
-        }
-
-        impl sea_orm::TryGetable for $name_id {
-            fn try_get(
-                res: &sea_orm::QueryResult,
-                pre: &str,
-                col: &str,
-            ) -> Result<Self, sea_orm::TryGetError> {
-                match String::try_get(res, pre, col) {
-                    Ok(v) => Ok($name_id(v)),
-                    Err(e) => Err(e),
-                }
-            }
-        }
-
-        impl sea_orm::sea_query::Nullable for $name_id {
-            fn null() -> sea_orm::Value {
-                sea_orm::Value::String(None)
-            }
-        }
-
-        impl sea_orm::sea_query::ValueType for $name_id {
-            fn try_from(v: sea_orm::Value) -> Result<Self, sea_orm::sea_query::ValueTypeErr> {
-                match v {
-                    sea_orm::Value::String(Some(x)) => Ok($name_id(*x)),
-                    _ => Err(sea_orm::sea_query::ValueTypeErr),
-                }
-            }
-
-            fn type_name() -> String {
-                stringify!($name_id).to_owned()
-            }
-
-            fn column_type() -> sea_orm::sea_query::ColumnType {
-                String::column_type()
-            }
-
-            fn array_type() -> sea_orm::sea_query::ArrayType {
-                String::array_type()
-            }
-        }
-
-        impl std::fmt::Display for $name_id {
-            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-                self.0.fmt(f)
-            }
-        }
-    };
-}
-
 macro_rules! create_id_type {
     ($name_id:ident, $key_prefix:literal) => {
-        string_wrapper!($name_id);
+        $crate::string_wrapper!($name_id);
 
         impl BaseId for $name_id {
             const PREFIX: &'static str = $key_prefix;
@@ -350,7 +284,7 @@ macro_rules! create_all_id_types {
         create_id_type!($name_id, $key_prefix);
 
         // Uid
-        string_wrapper!($name_uid);
+        $crate::string_wrapper!($name_uid);
 
         impl BaseUid for $name_uid {
             const ID_PREFIX: &'static str = $key_prefix;
